Basic UI design and components implemented
Created the basic UI using material-ui and in the process created components for the action buttons and the card in the center of the screen, probably not optimal as I am just learning React but it is working fine at the moment with multiple different image sizes in it. All components are 'started' in index.js > Main but Main doesn't deal directly with individual material-ui components such as RaisedButton or Card.