Star rating and event handling
Implemented star ratings that show at the bottom of the card, when the google maps api is integrated, it will be based on the rating grabbed from that but it essentially checks the rating, decides how many full star icons to use and whether or not a half star icon needs to be used. I also implemented a rudimentary event handling module that exports two functions for liking and disliking that are triggered when a button is pressed that corresponds to that action. The event handling module will talk with the animations module and trigger relevant animations when the events occur.