Card generation
Card generation now in place, has a few issues, mainly the size of images messing up the card by not staying within bounds, and going out of bounds due to the index used to traverse the nearby places array, but it works and grabs in the images for locations as well as the name and rating of that place. Also commented a bunch of the code.