handling errors using express
In the service there are no catch statements so if anything fails it will fall back to the catch in the controller, this also happens, in the case of getPlaceData service where the request doesn't _fail_ but it returns an invalid status code, in this case we throw the error ourselves. In the controller this is then caught in the catch block and passed into the express next() function which, by default, will send this to the client, to make it more readable we prepend the error returned with a predefined error message that is imported in the controller it will be used. In this case it is an apiGetRequestFailed error.